Stop passing a type parameter to IterableOnceOps#toSet
It should never be needed with a covariant Set. This means we could also
completely remove the type parameter from the definition of `toSet`,
however that would break source compatibility. This might be fixable by
adding a `toSet[B >: A]: Set[B]` as a deprecated extension method but I
haven't explored that.