Use applyOrElse for PartialFunctions
This avoids evaluating the guards twice. It can be significantly
(25% in my benchmarks) slower in the worst case (a PF with a single
clause with very efficient guard that usually does not match) but can
have much larger performance gains in other scenarios.

All flatMap-based implementations suitable for lazy collection types
are replaced by a new `View.Collect` primitive that uses an Iterator
instead, which is considerably faster.