Remove reflection-based class name computation for collections
This was applied inconsistently in the past. For example, `Map` used to
have an override that set the class name to “Map” in 2.12 which caused a
regression in SI-10932. Other basic collection types did not have such
an override so they got the reflective version, which led to issues
such as SI-9019.

This change removes the reflection-based name entirely. The basic types
`Iterable`, `Map`, `Set`, `Seq`, `LinearSeq`, `IndexedSeq`, `Buffer`,
`SortedSet`, `SortedMap` and `View` use their own name, concrete
implementations override it as before. User-defined collection types are
expected to override `className` if they do not want to use the default.

Small sets and maps now use the base name (“Set” or “Map”). This was
inconsistent in 2.13.0-M4 (sets used “Set” but maps used specific names
such as “Map.Map1”).

All user-extensible classes override `stringPrefix` instead of
`className` for compatibility with 2.12.