Support -Yimports for implicit preamble imports
Take `-Yimports:x,y,z` to mean `x,y,z` are root imports
of the form:
```
import x._ { import y._ { import z._ { code } } }
```
and where the import from a module (non-package) can
be disabled with usual idiom, an import at the package level.

The disabling import is registered when the import
is completed, so it must precede all usages.