CompileServer doesn't want to inherit I/O
The changes so ScriptRunner doesn't use the
CompileServer by default resulted in redirected
I/O hanging, for example: `fsc | more` would
not terminate after displaying output. This
was noticed by the smoke test, which greps
output.

When the fsc CompileClient is run, it will
spawn a CompileServer process. A previous
change to start the process with `inheritIO`
means the pipe client waits on output from
the child server process.

The CompileServer redirects its output to
log files using bear knives, so this commit
just removes the `inheritIO` configuration.
During a client session, output is sent to
the client via socket.