Support @Repeatable Java annotations.
Currently, duplicate classfile annotations cause a runtime crash
when the JVM sees them (due to a call to `getAnnotations` or the
like). Do instead exactly what Java does (since JEP-120): if the
annotation type is (meta-)annotated with `@Repeatable`, wrap the
annotations in an array and annotate the original element with a
new annotation of the type given by `Repeatable#value`.

It is now an error to have multiple annotations on the same tree
with the same `typeSymbol` if the symbol is a classfile annotation.

Fixes scala/bug#9529.