REPL prints long results
Previously, when REPL has many results
to output, it would print `a + b + etc`,
a potentially large expression tree.
`s"${a}${b}etc` is also a tree too large
for refchecks conversion; so just
use a string builder to build the
string to print, if there is more
than one result.