continue to make StringContext's checkLengths available to clients
partially reverts #6774. I found in the community build that at least
one string interpolator in the wild (hello lightbend-emoji!) was using
this method -- because it's rather useful, so actually we should
continue to provide it.

I agree with the part where we moved it to the companion object, so I
kept that, but I made it public there.

and I left the old one around as deprecated, to facilitate
cross-building.  that means we still have a "funny interpolator"
around for one more cycle, but oh well.