Java annotations are not classes.
Previously, `JavaParsers` and `ClassfileParser` would "fix-up"
Java annotations to be non-abstract classes with `Annotation`
and `StaticAnnotation` as parents. This caused all manner of
strangeness and invalid classfiles (see linked tickets) and
was surprising, to say the least.

Now, the only special dispensation given to Java annotations
is that they get a public, no-args class constructor, because
that's what it takes to get `typedAnnotation` to be able to
look up the type. This means that `new Foo`, where `Foo` is
a Java annotation, still compiles (as it does right now,
actually). Part of the reason for this commit is to provoke
a discussion about how to fix that.

Fixes scala/bug#8778.
Fixes scala/bug#9400.
Fixes scala/bug#9644.