Avoid ScalaRuntime for generic arrays that are refined to concrete arrays
The value bound by a type pattern takes on an intersection type of the
scrutinee type and the pattern type. In the enclosed test, this leads
to a type like `Array[T] with Array[AnyRef]`.

This is erased to `Array[Object]`, thanks to use of `intersectionDominator`
in `ErasureMap`. However, some related code in the erasure tree transform
was unable to see through this refinment type and emitted the slower
code for array access using runtime type checks in ScalaRuntime.