Lazier LazyList.
As lamented in scala/bug#10696 and bemoaned in scala/collection-strawman#367,
`LazyList` (and `Stream` before it) does not have a way of representing a
collection with uncomputed emptiness. This adds a third subclass of `LazyList`,
`LazyList.Suspended`, which wraps a closure returning a `LazyList`, and only
evaluates it when needed. This is about as lazy of a list as I can imagine, now.

Fixes scala/bug#10696.
Fixes scala/collection-strawman#367.