Make quasiquote pattern matching deterministic
As per the discussion at https://github.com/scala/scala/commit/7184fe0d3740ac8558067c18bdf449a65a8a26b9#r29651930,
all we want to avoid is name collision among the holes in a single
pattern. For that, `c.freshName(<a constant>)` itself is sufficient
and the randomness is not needed. This is what scala.meta does,
and works just as well.

This fixes https://github.com/scala/bug/issues/11008