Enhanced the right set of type bounds
In inferMethodInstance a new list of undetermined symbols is computed
and returned. However, the bounds enhancement introduced in #6140 is
applied to the list of previously undetermined symbols, rather than the
new list. This commit applies the bounds enhancement to the newly
computed list.

None of the other callsites of enhanceBounds appear to have the same
issue.

Fixes scala/bug#10758.