Respect BaseTypeSeq invar (3) symbols are distinct
In a BTS, a singleton type is redundant with its underlying type.
In other words, what could we learn about the superclasses of
a singleton type that is not captured entirely by its super type?

This duplicate type symbol leads to confusion during asSeenFrom
when looking up the base class that defines the T type param
(Since `typeOf[this.type].typeSymbol == typeOf[Poly[_]].typeSymbol`,
 we return `this.type` for the base type at Poly,
 which does not have the expected type params.)

The interesting part is that you have to embed the singleton in
a compound type to trigger the bug because SubType
(a supertype of SingletonType) delegates `baseType` to its
underlying (super) type!

TODO: a similar deviation still exists in AbstractTypeRef.
I hesitate to bring that one in line with the spec, because I
could imagine it actually being used.