Remove call to Reverse.
The code fos suspendingTypeVars only cares about the presence of all
the variables, not the orden between them. Yet, the `typeVarsInType`
function was running a `reverse`, which spends twice as much memory.
We create a special function

We do not delete the initial function because it is a public method.