Add a default to ImplicitSearch ctor arg to unbreak shapeless
The introduction of the isByNamePt constructor argument of
ImplicitSearch breaks the use of this constructor by shapeless. This
commit fixes that by adding a default argument.

I'd be happy to also reorder the arguments so that isByNamePt comes
last, or alternatively reinstate the previous constructor signature via
a secondary constructor.