Retain empty parens as unit value
When parsing infix expression, a Parens node
is created for parens. If it is empty, they
wrote unit value. Otherwise, per the spec,
the possibly many comma-separated arguments
are taken as arguments.

`() == ()` is now what it is.

`x op()` is corrected to `x.op()`. The dot-averse
must rely on postfix and empty application
to use `x op`.

Restore wording adapt is removed.