Accept numeric literal separators
The separator character is underscore
as in Java.  They may be mixed
arbitrarily in a numeric literal, but
only in the interior: the literal, and
each part of double, must begin and
end with a digit.

Add spec for underscore numeric separator.

Test literal positions.

Verifying visually that literals with
underscores are the correct length.