WIP Remove some statistics
Profiles suggest that in very hot code paths, there is still some
overhead from checking if statistics are enabled, even with the
invokedynamic tricks we're using.

These are the spots I see the most overhead.

I'm interested to see if we measure a performance improvement
on our benchmark server after this change.