Remove some statistics
Profiles suggest that in very hot code paths, there is still some
overhead from checking if statistics are enabled, even with the
invokedynamic tricks we're using.

These are the spots I see the most overhead.

This may be limited to JDK 8, but we still have plenty of users
on that release. Or it maybe a bug with Java Flight Recorder,
(again, perhaps just on some JDK version), that is giving me
false reports.