Traverse synthetic top-level members
The previous commit was a good fix for the wrong problem. After some
more investigation, the problem at hand is that we don't traverse
synthetic top-level trees in `API` and therefore we don't extract the
used names from them.

This was causing us to ignore any names in the synthetic companion and,
on top of that, also synthesized top-level members like package object
definitions! Now, thanks to removing the filtering out that was
happening in `isTopLevel`, our previous issue is fixed and with it also
a pending test checking that package objects are recognized when
introduced in a change `packageobject-and-traits`.

This commit also adds a `checkRecompilations 2` in one of the tests
related to package objects to make sure that test is behaving correctly.

Rewritten from sbt/zinc@663ec5ab53d733069f1854839aa387bba9f97238