Don't re-apply type maps to info of unchanged symbols
The fast path in `mapOver` would avoid cloning symbols if the map
had no effect on any of their infos. However, if a symbol with a
changed info was found, it would use `cloneSymbolsAndModify` to
apply itself to a fresh clone of those symbols, with the result
that it would re-apply itself to the infos of symbols clone from
symbols that were unchanged.

This avoids that.

In library/reflect/compiler, this avoids 51358 _direct_ repeated
calls to `TypeMap#apply`. Since those may go off and do more work,
it's more than that.