Uncurry of Java varargs respects Java's universal trait nescience
If `U` is a universal trait (i.e., one extending from `Any` rather
than `AnyRef`), Java still sees it as a normal interface, and in
a method with a generic varargs parameter `<T extends U>`, interprets
that as declaring an array argument `U[]` after erasure. However,
due to a pre-value-class implementation of `isUnboundedGeneric`,
such a type parameter was considered unbounded, and therefore Scala
considered its erasure to be `Object[]` instead. This causes a runtime
`NoSuchMethodError`, of course.

I moved `isUnboundedGeneric` from `Types` into `UnCurry`, since that's
the only place it was used, and its proximity to `isBoundedGeneric`,
which is neither defined as nor synonymous with `!isUnboundedGeneric`
appeared likely to cause confusion. (_That_ method is only used in
`SpecialiseTypes`, but I didn't want to change an unrelated file.)

Running into this bug is probably penance for using Hibernate with
Scala, but I promise I'm "just following orders".

Fixes scala/bug#11109.