Add UpperBound and LowerBound methods to Type
This commit adds to the "Type" abstract class two methods that return
the upper and lower type bound of the type in question.

It was noted that the call to `Type.bounds` was a memory hotspot.
Looking at the code, we noted that a lot of times, the call
to `.bounds` was immediately followed by a call to `.hi` or `.lo`,
i.e., only one of the bounds was needed at each point.
This inspired us to add methods that would get each bound directly.