Take omitted Unit for procedure syntax
The weakness of procedure syntax was that
it is easy to accidentally leave out an
equals sign, especially as a beginner.

The strength of procedure syntax was that
it was a strong visual signal for an
experienced eye that the method is Unit-valued.

This commit takes an omitted result type
as explicit Unit: `def f(): = ()`.

The extra colon over an inferred result type
means that it is not an easy beginner trap,
but it is also pithy and easy to spot for
experienced eyes.

For aesthetics, also accept `:=` for this
idiom, since an identifier is not otherwise
valid in this position.

This identifier was once proposed as
the Ur-assignment operator, so this syntax
has the added benefit of triggering one's
spidey sense, or Speyside sense, depending
on the time of day. It is also familiar
from sbt in a different usage.