Cache results of `findAssociatedFile` for performance
We want to make sure that if this is ever run, we don't repeat the
computation for the same class over and over again across the same
compiler run. So we install a per-cache run.

Rewritten from sbt/zinc@ec0e2b3afe91af378553cb5a6b78c289994dc52c