Replace a list of booleans by a mutable BitSet.
The modified code saves, for each element of a list, a boolean state
variable that is modified before an operation and restored afterwards.
The code was using a `List.map`, which creates a linked list with the
same length as the input list, with as many Boolean objects.
We change the code to use a BitSet instead, which needs less memory.