A quick and dirty fix for Scala Bug 10896
This commit addresses https://github.com/scala/bug, number 10896.

The bug was manifesting itself in a set of mutually recursive calls to
methods `normalize`, `normalizeImpl`, and `betaReduce`, none of which
was modifying the actual method they act upon. This resulted in an
infinite recursion, leading to a StackOverflow Error.

The quick fix in this commit simply checks if the `betaReduce` method
has modified the object. If there is no change, then we short-circuit
with a generic `MalformedType ` error.
This fixes the case in the bug, but may not be general enough.
In Dotty, it uses a more general solution by means of Kind-Checking.