Attend Code Review Comments.
This incorporates some changes, suggested by Adrian Moors, to use more
efficient data-structures (in-place), with some modifications.
We use two additional temporary lists, `tparamsBord1` and `bord2`, to
keep a border of the transitive closure: the list of elements
introduced by the previous iteration. This avoids repeating
`Type.contains` computations between same pairs.

We add to the `Collections` a function to partition a list into two ListBuffers.