Remove analyzer caching
It hurts performance. Analyses are really memory-intensive. Caching
them uses a lot of memory, and basically defers collecting them from
minor to major GC (they are stored using SoftReference).
With a 1.5G heap, the JVM is basically doing major GC all the time
when running the optimizer, which drasitcally hurts performance. With
a 3G heap this looks better. However, even with a 3G heap the overall
performance is worse than without any caching on a 1.5G heap.

Generally, the optimizer has a huge amount of memory churn, this can
be observed easily with a profiler. The best way to improve on that is
probably to run fewer analyzers, i.e., to fuse local optimizations that
use the same kind of analysis.