Don't reject views with result types which are TypeVars
On the matchesPtInst fast path views are pruned without first being
applied. This can result in a false negative in HasMethodMatching if the
view has a result type which is a not fully instantiated TypeVar. The
fix is to fall back to the slow path in that case.

Fixes scala/bug#11174.