JvmWiseLUB: Use Bloom Filter for quick checks
The `jvmWiseLUB` uses a method `firstCommonSuffix`, which takes two
lists of ordered subclasses and returns the first element that is in
both lists. This method takes quadratic time.

To try to improve its performance, we add a little bloom filter for
each list. At the beginning we compute the bloom filter for each list,
and afterwards we first check in the filter before looking through the list.
Our "bloom filter" is rather simple: a Long variable, which has 64 bits.
We use a simple hash which maps each string to a bit position between 0 and 63.

The benefit is that computing the filter takes linear time, with one
run through the loop for each list, but checking if an element appears
in the filter only takes constant time (computing the "hash").