Split TypeMaps with Variance
We noted that the "trackVariance" attribute of a type map was never
changed, which suggests that it was a static element better expressed
through a type than through a boolean field.

We modify the TypeMap class Hierarchy, so that type maps that do need
to track variance have their own specific  class "VariancedTypeMap".
We move the variance-handling methods, and branches of methods, there.

We add a companion object to VariancedTypeMap, which takes care of
checking if a TypeMap is a VariancedTypeMap or not.

This code seems to break the execution in partest of one test,
for which we are not sure if it is actually fixing a bug.