Split TypeMaps with Variance
We noted that the "trackVariance" attribute of a type map was never
changed, which suggests that it was a static element better expressed
through a type than through a boolean field. For this reason,
we modify the TypeMap class Hierarchy, so that type maps that do need
to track variance have their own specific  class "VariancedTypeMap".

We move the variance-handling methods, and the branches of some
methods, from the `TypeMap` to the `VariancedTypeMap` class.
This also allows us to replace some calls to `mapOverArgs`, which zips
over two lists, with a call to a `mapConserve`.