Don't lift try blocks into separate methods in boolean expressions
Boolean not/and/or are compiled to conditional jumps, so a `try-catch`
expression in argument position can be emitted directly, there is no
need to it into a synthetic method.

In general, `try-catch` expressions in argument position are emitted
in separate methods because the JVM clears the call stack when entering
a handler. This would discard stack values that are still used, e.g.,
the receiver of an invocation.