Performance tweak: avoid traversals and shortcut.
The `contains` and `find` operations, only need to find the first
element that is equal / satisfies the property, and end immediately.
They do not modify the type at all, so no full visit is needed.

We modify the implementations of these methods, to avoid traversals.

We modify the `FindTreeTraverser` to incorporate the same kind of
short-cut that we have used in the `ContainsTypeTraverser`.
We modify our `ContainsTypeTraverser` to use a `FindTreeTraverser`,
instead of calling `forach` (which created one traverser every time).