When parsing a Scaladoc table do not consume non-table content
Fixes two issues by reducing greediness of Scaladoc table parsing,

- Content appearing directly after table ends is not lost in the call to blockEnded("table")
- Cell content breaking over more than one line is not parsed

By looking further ahead and ensuring the next line of content is a valid table row both these issues are avoided
and subsequent parsing is simplified because the rows being parsing have a known structure.

This approach preserves the fast failing check on existence of a new table row via the same quick check on the
next char as the previous non-regex approach.

Scaladoc tables are a subset of GFM tables. One restriction over GFM is the requirement for table rows to start
and end with the | character with no leading/trailing whitespace. The test for leading whitespace not being
ignored is turned on and confirms the restriction is active.