[no-merge] Make xml pos consistent with scanner after resume
XML parser uses current offset to compute end of
token offset, which is wrong. This commit makes
sure at least to back up to the CR of a line
ending so that the position is contained by the
position used by scanner after scanner.resume.
nextToken adjusts the position of the line ending.