More thorough treatment of @strictfp.
(Just getting my toes wet again.)

The JVM accepts `ACC_STRICT` as a modifier flag on non-abstract
methods only. Java itself accepts `strictfp` on both methods
(in which case it applies to the method) and classes (in which
case it applies to all methods within the class, as well as
nested and inner classes thereof.) Scala has somewhat more ways
of nesting methods and classes than Java, so I've extrapolated
the rule to be: methods nested inside of a class/module/method
definition marked `@strictfp` are strict.

I also fixed the interaction with value classes: when a method
body on a value class was moved to the companion, its `@strictfp`
attribute wasn't moved along with it.

The test case covers nested/inner/local classes and methods, as
well as extension methods. I'm leaving specialization to the
existing specialization+strictfp tests.

Fixes scala/bug#7954.