Handle bad process start and lazily
Improve handling of failure of a process
to start. In a `CompoundProcess`, catch
exceptions and set a special exit value.
For `lazyLines`, also supply a value so
that requesting thread doesn't hang.

It would be nicer to stow the exception
and rethrow instead of using special
integers.

Test for stale thread.