Align detection of implicit recursion with the SIP
The by-name implicits SIP specifies that a recursive occurence of an
implicit should tie the knot if there is at least one by-name parameter
between the outermost and the innermost occurrence. Prior to this commit
the implementation required that either the innermost or the outermost
occurrence must be by-name. This commit brings the implementation into
line with the SIP.