Fix interactions between macros and by-name implicits
While prototyping a version of shapeless which replaces Lazy with
by-name implicits it emerged that implicit resolutions which involve
both by-name implicits and implicit macros can produce dictionaries
which type checked but which had broken owner chains.

This had been anticipated (see 5dd99093) but the ownership change
performed there wasn't thoroughgoing enough.

The implicit dictionary is now encoded as,

  final class LazyDefns$n extends java.io.Serializable {
    final val rec$m ...
  }
  val lazyDefns$n = new LazyDefns$
  lazyDefns$n.rec$m

rather than as an object, to eliminate the unecessary overhead of object
lazy initialization and to allow serialzation in rare cases where the
enclosing value is captured.

With this commit a Lazy-free shapeless build is possible.