Fix crasher regression with implicit classes and default params
Since the changes to make the compiler output deterministic, default getter
symbols must be entered eagerly before the trees are created. This happens
in `enterDefDef`, but that method is bypassed when entering the synthetic
symbol for an implicit class factory method.

This commit enters the default getter symbols in this case, as well, avoiding
a later crash.