[nomerge] No static forwarders for bridges implementing abstract methods
In 2.12.7, #7035 added the `bridge` flag to static forwarders that are
generated for bridge methods. (2.13 geneartes no forwarders for bridges,
but we wanted to stay binary compatible in 2.12.)

Unfortunately the change caused even more bridges to be generated,
namely for bridge methods that implement an abstract member. Now we
exclude them again, which brings the binary interface back to the state
of 2.12.6.

Fixes scala/bug#11207