Reset access boundary of companion module symbols
When a synthetic module symbol was created with `ensureCompanionObject`
during `enterClassDef`, _before_ an explicitly written companion module
that appears later in the list of statements, the module symbol initially
has the access boundary of the class (as copied by `caseModuleDef`).

When the module is later entered, this symbol is mutated to reflect
what is written in source code:

https://github.com/scala/scala/blob/bdfc43b3e0b3e34d76ee9bd248b6fb49f38e71eb/src/compiler/scala/tools/nsc/typechecker/Namers.scala#L467-L470

This code fails to remove the access boundary.