Fix ArrayOps.sorted and add tests for ArraySeq sorting
These are the missing tests for https://github.com/scala/bug/issues/11187.
The other sort methods show no remaining instances of the bug but there
was an unrelated but uncovered by the tests.

A different fix (#7480) was merged in the meantime but it would be
preferable to use the JUnit tests here instead of the partest in
#7480.