traverse does not take an IterableOnce of Futures
Hello, I just want to draw your attention to the fact that the current description of the 'in' parameter seems incorrect. The most obvious  symptom of this is the fact that the description is identical to that of the 'in' parameter of the sequence method, but while the 'in' parameter of sequence is an `IterableOnce` of Futures, the 'in' parameter of traverse is not an `IterableOnce` of Futures, rather, its elements will be mapped to Futures by invoking the function that is the 'fn' parameter. I am not suggesting that my changes be made verbatim, you'll certainly do a much better job of figuring out an 'appropriate' way of correcting the description (e.g. one consistent with the rest of the Scaladoc).