Add proper type constraints for CC type parameters
This also provides a nice solution for AnyConstr (see
https://github.com/scala/scala/pull/7077 and
https://github.com/scala/scala/pull/7072): We simply don’t need it
anymore.

We really want to constrain both IterableCC and MapCC to Iterable
(the latter because it’s needed for MapView) but that causes collisions
after erasure, so we have to go back to IterableOnce for IterableCC.
Using something more specific (like MapOps with Iterable) for
MapCC is not possible because in MapView the MapCC is not actually
MapView but only View.