VectorBuilder avoids copying data if possible
When calling VectorBuilder#addAll(Vector), if the builder is currently
empty, it will share structure with the argument Vector in order to
avoid iterating and copying the entire argument Vector.

This is very helpful especially for the Vector#concat implementation,
which in most cases will work by creating a new builder, and adding
both vectors to the builder.