Fix rvm in Travis config
An "rvm" entry, here "2.2", in a Travis configuration file without a
"language: ruby" declaration seems to silently ignore the former.
Travis doesn't support multiple languages, so it seems manual
installation of a ruby version with rvm is necessary.

Diagnostic commands for ruby have been added to help avoid this
problem in the future.