Deconst result type of non-getter method application
When a constant value is returned from a method --
even one without argument list -- we must deconst,
as we know nothing about the method's purity, and
thus cannot constant fold through that application.

By analogy to the nullary method case in adapt, deconst
result, unless it's a stable method invoked on safe to inline expr

Also, deconst result type of getters during fields phase:
constant folding time is over (that happens before, during typers),
but constant types have not yet been eliminated (that happens during erasure).

So, we must deconst the result type of methods to avoid treating
their application as a constant...

In patmat, deconst info of default label for collapsed switch

Now that we're more conservative about deconsting
the result of a method application (nullary or not),
a bug in the pattern matcher is revealed: before,
we would constant fold the jump to the default label,
as well as the labeldef itself (since its type wasn't deconsted).
but its disappearance was compensated by the disappearance
of the jump to it.

Now, the jump isn't deconsted, and neither is the label def.

See scala/bug 10788 and 10768.